requirements.txt
Django==2.1
fabric==2.4.0
gunicorn==19.9.0
ipython==5.8.0
psycopg2==2.7.5
django-phonenumber-field==2.1.0
phonenumberslite==8.9.16
djangorestframework==3.9.0
django-extensions==2.1.3
django-allauth[with_social]==0.38.0
django-rest-auth==0.9.3
django-bootstrap-form==3.4
django-localflavor==2.1
Pillow==5.3.0
django-cors-headers==2.4.0
django-anymail[amazon_ses]
celery[redis]
django-celery-results==1.0.4
redis==3.0.1
django-nose==1.4.6
model-mommy==1.6.0
##Till 0.7.0 is released, compatible with Django 2.1
##https://github.com/justquick/django-activity-stream/issues/401
git+https://github.com/justquick/django-activity-stream.git
django-filter==2.0.0
XlsxWriter==1.1.2
names==0.3.0
simple-salesforce==0.74.2
django-simple-history==2.6.0
django-jsonfield-compat==0.4.4
channels==2.1.6
django-rest-swagger==2.2.0
channels-redis==2.3.3
service-identity==18.1.0
daphne==2.2.4
googlemaps==3.0.2
lxml==4.3.0
bs4==0.0.1
pygeocodio==0.10.0
sendgrid==5.6.0
